Bluetooth: Mesh: Implement functions for the EMDS.
Add implementation for the Emergency Data Storage. This storage should
be have its own storage area that can be quickly stored to at when
necessary. This way we can store data that is changing a lot without
wearing out the flash, and have a predefined time for storing the data.
There should be possible to determine how mush time is needed to store
all data configured to be stored in this area.

Signed-off-by: Ingar Kulbrandstad <ingar.kulbrandstad@nordicsemi.no>